Created Controllers folders & seperated each endpoints into seperate file
