+++ server.js
@@ -3,6 +3,11 @@ const bcrypt = require("bcrypt-nodejs");
 const cors = require("cors");
 const knex = require("knex");
 
+const register = require("./Controllers/register");
+const signIn = require("./Controllers/signIn");
+const profile = require("./Controllers/profile");
+const image = require("./Controllers/image");
+
 const db = knex({
   client: "pg",
   connection: {
@@ -24,129 +29,30 @@ const app = express();
 app.use(cors());
 app.use(express.json());
 
-const database = {
-  users: [
-    {
-      id: "123",
-      name: "john",
-      email: "john@gmail.com",
-      password: "cookies",
-      entries: 0,
-      joined: new Date(),
-    },
-    {
-      id: "124",
-      name: "sally",
-      email: "sally@gmail.com",
-      password: "bananas",
-      entries: 0,
-      joined: new Date(),
-    },
-  ],
-  login: [
-    {
-      id: "987",
-      hash: "",
-      email: "john@gmail.com",
-    },
-  ],
-};
-
 app.get("/", (req, res) => {
   res.send(database.users);
 });
 
 /* SignIn*/
 app.post("/signin", (req, res) => {
-  // Load hash from your password DB.
-
-  /* instead of using res.send() we can use res.json(). express  comes with a 
-     build in JSON method on response that we can use and it has some added
-     features when responding with json
-    */
-  // res.send("signinn");
-
-  db.select("email", "hash")
-    .from("login")
-    .where("email", "=", req.body.email)
-    .then((data) => {
-      const isValid = bcrypt.compareSync(req.body.password, data[0].hash); // true
-      if (isValid) {
-        return db
-          .select("*")
-          .from("users")
-          .where("email", "=", req.body.email)
-          .then((user) => res.json(user[0]))
-          .catch((err) => res.status(400).json("Unable to get user"));
-      } else {
-        res.status(400).json("Wrong credentials");
-      }
-    })
-    .catch((err) => res.status(400).json("Wrong credentials"));
+  signIn.handleSignIn(req, res, bcrypt, db);
 });
 
 /*Register*/
 
 app.post("/register", (req, res) => {
-  const { name, email, password } = req.body;
-  const hash = bcrypt.hashSync(password);
-  db.transaction((trx) => {
-    trx
-      .insert({
-        hash,
-        email,
-      })
-      .into("login")
-      .returning("email")
-      .then((loginEmail) => {
-        return trx("users")
-          .returning("*")
-          .insert({
-            name,
-            email: loginEmail[0],
-            joined: new Date(),
-          })
-          .then((user) => {
-            res.json(user[0]);
-          });
-      })
-      .then(trx.commit)
-      .catch(trx.rollback);
-  }).catch((err) => {
-    res.status(400).json("unable to register");
-  });
+  register.handleRegister(req, res, bcrypt, db);
 });
 
 /* user home profile */
 app.get("/profile/:id", (req, res) => {
-  const { id } = req.params;
-  db.select("*")
-    .from("users")
-    .where({
-      id: id,
-    })
-    .then((user) => {
-      if (user.length) {
-        res.json(user[0]);
-      } else {
-        throw error;
-      }
-    })
-    .catch((err) => {
-      res.status(404).json("user not found");
-    });
+  profile.handleProfileGet(req, res, db);
 });
 
 /* image endpoint */
 
 app.put("/image", (req, res) => {
-  const { id } = req.body;
-  db("users")
-    .where("id", "=", id)
-    .increment("entries", 1)
-    .returning("entries")
-    .then((entries) => res.json(entries[0]))
-    .catch((err) => res.status(400).json("unable to get entries"));
+  image.handleImage(req, res, db);
 });
 
 app.listen(3000, () => {
